0_Summary/obsidian_summariser/pyproject.toml
[project]
name = "obsidian_summariser"
version = "0.1.0"
description = "Obsidian × Git × LLM auto-summary"
authors = [{ name = "austin-kim", email = "kmaestro043@gmail.com" }]
readme = "README.md"
requires-python = ">=3.11"

dependencies = [
  "openai>=1.25.0",
  "tenacity>=8.2.3",
  "click>=8.1.7",
  "python-dotenv>=1.0.1",
  "jinja2>=3.1.3",
  "tiktoken>=0.6.0"
]

[project.scripts]
obsidian_summariser = "obsidian_summariser.cli:app"

[tool.poetry]
packages = [{ include = "obsidian_summariser" }]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"


[tool.poetry.group.dev.dependencies]
pytest = "^8.4.0"
ruff = "^0.11.13"
black = "^25.1.0"
isort = "^6.0.1"
pre-commit = "^4.2.0"
